исправления по комментариям
